test: Allow use of FsHelper without a config

See merge request u-boot/u-boot!141

+++ test/py/tests/fs_helper.py
@@ -7,6 +7,7 @@
 
 import re
 import os
+import shutil
 from subprocess import call, check_call, check_output, CalledProcessError
 from subprocess import DEVNULL
 import tempfile
@@ -63,16 +64,21 @@ class FsHelper:
 
         # Use a default filename; the caller can adjust it
         leaf = f'{prefix}.{fs_type}.img'
-        self.fs_img = os.path.join(config.persistent_data_dir, leaf)
+        if config:
+            self.fs_img = os.path.join(config.persistent_data_dir, leaf)
+            if os.path.exists(self.fs_img):
+                os.remove(self.fs_img)
+        else:
+            self.fs_img = leaf
 
         # Some distributions do not add /sbin to the default PATH, where mkfs
         # lives
         if '/sbin' not in os.environ["PATH"].split(os.pathsep):
             os.environ["PATH"] += os.pathsep + '/sbin'
 
-        self.srcdir = None
         self.tmpdir = None
-        self._do_cleanup = True
+        self.srcdir = None
+        self._do_cleanup = False
 
     def _get_fs_args(self):
         """Get the mkfs options and program to use
@@ -108,38 +114,43 @@ class FsHelper:
         fs_img = self.fs_img
         with open(fs_img, 'wb') as fsi:
             fsi.truncate(self.size_mb << 20)
+        self._do_cleanup = True
 
-        try:
-            mkfs_opt, fs_lnxtype = self._get_fs_args()
-            check_call(f'mkfs.{fs_lnxtype} {mkfs_opt} {fs_img}', shell=True,
-                    stdout=DEVNULL if self.quiet else None)
+        mkfs_opt, fs_lnxtype = self._get_fs_args()
+        check_call(f'mkfs.{fs_lnxtype} {mkfs_opt} {fs_img}', shell=True,
+                stdout=DEVNULL if self.quiet else None)
 
-            if self.fs_type.startswith('ext'):
-                sb_content = check_output(f'tune2fs -l {fs_img}',
-                                        shell=True).decode()
-                if 'metadata_csum' in sb_content:
-                    check_call(f'tune2fs -O ^metadata_csum {fs_img}', shell=True)
-            elif fs_lnxtype == 'exfat':
-                check_call(f'fattools cp {self.srcdir}/* {fs_img}', shell=True)
-            elif self.srcdir and os.listdir(self.srcdir):
-                flags = f"-smpQ{'' if self.quiet else 'v'}"
-                check_call(f'mcopy -i {fs_img} {flags} {self.srcdir}/* ::/',
-                        shell=True)
-        except CalledProcessError:
-            os.remove(fs_img)
-            raise
+        if self.fs_type.startswith('ext'):
+            sb_content = check_output(f'tune2fs -l {fs_img}',
+                                    shell=True).decode()
+            if 'metadata_csum' in sb_content:
+                check_call(f'tune2fs -O ^metadata_csum {fs_img}', shell=True)
+        elif fs_lnxtype == 'exfat':
+            check_call(f'fattools cp {self.srcdir}/* {fs_img}', shell=True)
+        elif self.srcdir and os.listdir(self.srcdir):
+            flags = f"-smpQ{'' if self.quiet else 'v'}"
+            check_call(f'mcopy -i {fs_img} {flags} {self.srcdir}/* ::/',
+                    shell=True)
 
     def setup(self):
         """Set up the srcdir ready to receive files"""
         if not self.srcdir:
-            self.tmpdir = tempfile.TemporaryDirectory('fs_helper')
-            self.srcdir = self.tmpdir.name
+            if self.config:
+                self.srcdir = os.path.join(self.config.persistent_data_dir,
+                                           f'{self.prefix}.{self.fs_type}.tmp')
+                if os.path.exists(self.srcdir):
+                    shutil.rmtree(self.srcdir)
+                os.mkdir(self.srcdir)
+            else:
+                self.tmpdir = tempfile.TemporaryDirectory('fs_helper')
+                self.srcdir = self.tmpdir.name
 
     def cleanup(self):
-        """Remove created files"""
+        """Remove created image"""
         if self.tmpdir:
             self.tmpdir.cleanup()
-        os.remove(self.fs_img)
+        if self._do_cleanup:
+            os.remove(self.fs_img)
 
     def __enter__(self):
         self.setup()
@@ -188,6 +199,8 @@ class DiskHelper:
         self.fs_list = []
         self.fname = os.path.join('' if cur_dir else config.persistent_data_dir,
                                   f'{prefix}{devnum}.img')
+        self.remove_img = True
+        self._do_cleanup = False
 
     def add_fs(self, fs_img, part_type, bootable=False):
         """Add a new filesystem
@@ -217,6 +230,7 @@ class DiskHelper:
         img_size = pos
         try:
             check_call(f'qemu-img create {self.fname} {img_size}M', shell=True)
+            self._do_cleanup = True
             check_call(f'printf "{spec}" | sfdisk {self.fname}', shell=True)
         except CalledProcessError:
             os.remove(self.fname)
@@ -230,15 +244,17 @@ class DiskHelper:
             pos += fsi.size_mb
         return self.fname
 
-    def cleanup(self, remove_full_img=False):
+    def cleanup(self):
         """Remove created file"""
-        os.remove(self.fname)
+        if self._do_cleanup:
+            os.remove(self.fname)
 
     def __enter__(self):
         return self
 
     def __exit__(self, extype, value, traceback):
-        self.cleanup()
+        if self.remove_img:
+            self.cleanup()
 
 
 def mk_fs(config, fs_type, size, prefix, src_dir=None, size_gran = 0x100000,
